Refactor to use Go 1.10's net.Pipe code
This patch backports the Go 1.10 net.Pipe code so that it can be used to
provide deadline semantics for versions of Go earlier than 1.10. The
MemConn package's previous implementation of these semantics was
incomplete, and instead of trying to simply enhance the implementation
to match the one in Go 1.10, it made sense simply to backport the one
from Go 1.10 instead.

Additionally, the package's functions and the functions of the Provider
type now match the signatures found in the "net" package re: the Listen
and Dial variants that directly accept "net.Addr" types.