Optimize the listener close process
This patch refactors how listeners are closed with relationship to
pending I/O. While the provider must track connections in order to
prevent a duplicate address from listening on a specified network, the
remote endpoints that spawn from these connections need not have unique
addresses. To this end a listener no longer cares about tracking active
connections and the connection doesn't need to remove itself from the
tracker. Instead an event system with a done channel is used so that
when a listener is closed all of its pending I/O is unblocked and closed
immediately.